# chore: enable simd optimizations for aarch64