for ascii pack and unpack,

Also optimize scalar unpack for both x86 and aarch64.
We had to fix the bug in https://github.com/dragonflydb/dragonfly/pull/5140 and now we load chunks of 7 bytes
during unpacking. This greatly degraded the performance of
scalar unpack, so we now use the "naive" byte by byte implementation
which actually faster then using 7-byte loads on both x86 and aarch64.

On c4a (aarch64):

Benchmarks before:
```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
BM_PackNaive             222 ns          222 ns     18936335
BM_Pack                  222 ns          222 ns     18956309
BM_Pack2                 222 ns          222 ns     18951694
BM_PackSimd              220 ns          220 ns     19103906
BM_PackSimd2             223 ns          223 ns     18861252
BM_UnpackNaive           229 ns          229 ns     18228081
BM_Unpack                743 ns          743 ns      5643824
BM_UnpackSimd            744 ns          744 ns      5648469
```

Benchmarks after:
```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
BM_PackNaive             221 ns          221 ns     18971332
BM_Pack                  222 ns          221 ns     18963948
BM_PackSimd             97.2 ns         97.2 ns     43226095
BM_PackSimd2            96.6 ns         96.6 ns     43491371
BM_Unpack                228 ns          228 ns     18397585
BM_UnpackSimd            101 ns          101 ns     41733901

```

We improved scalar unpack by x3 from 743ns to 228ns, and improved vectorized unpack by x7.
We improved vectorized pack by x2.


On graviton2 (c6g):

before:
```
BM_PackNaive             465 ns          465 ns      1507214
BM_Pack                  434 ns          434 ns      1613102
BM_Pack2                 434 ns          434 ns      1613125
BM_PackSimd              434 ns          434 ns      1612817
BM_PackSimd2             434 ns          434 ns      1612666
BM_UnpackNaive           519 ns          519 ns      1347286
BM_Unpack                927 ns          927 ns       755409
BM_UnpackSimd            929 ns          929 ns       753418
```

after:

```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
BM_PackNaive             463 ns          463 ns      1511640
BM_Pack                  434 ns          434 ns      1612481
BM_PackSimd              241 ns          241 ns      2912345
BM_PackSimd2             239 ns          238 ns      2940373
BM_Unpack                515 ns          514 ns      1362422
BM_UnpackSimd            271 ns          271 ns      2581556
```

BM_PackSimd twice faster, BM_UnpackSimd x3.5 faster.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>
<!--
**Commits Must Be Signed and Your PR title must conform to the conventional commit spec**
  * See: https://github.com/dragonflydb/dragonfly/blob/main/CONTRIBUTING.md
  * Please follow the section on `pre-commit hooks`, a linter will validate before you push

  Example PR Title: <type>(<scope>)!: <description>

  * `type` = bug, chore, feat, fix, docs, build, style, refactor, perf, test
  * `!` = OPTIONAL: signals a breaking change
  * `scope` = Optional when `type` is "chore" or "docs"
  * `description` = short description of the change

Examples:

  * chore(examples): Clarify `docker` usage #120
  * docs(readme): Fix Example Links #121
  * feat(ingest)!: Add new ingest #122
  * fix(ingest): Refactor for loop to list comprehension #123
-->